// microservices/orders/src/order-part-request/entities/enums/part-request-status.enum.ts
// microservices/orders/src/order-part-requests/entities/enums/part-request-status.enum.ts

export enum PartRequestStatus {
    SOLICITADO = 'SOLICITADO',
    EN_BUSQUEDA = 'EN_BUSQUEDA',
    ENCONTRADO_NACIONAL = 'ENCONTRADO_NACIONAL',
    BUSQUEDA_INTERNACIONAL = 'BUSQUEDA_INTERNACIONAL',
    GESTION_INTERNACIONAL = 'GESTION_INTERNACIONAL',
    ESPERA_DE_PAGO = 'ESPERA_DE_PAGO',
    EN_PROCESO_DE_PEDIDO = 'EN_PROCESO_DE_PEDIDO',
    LLEGADO_ASIGNADO = 'LLEGADO_ASIGNADO',
    PENDIENTE_INGRESO = 'PENDIENTE_INGRESO',
    INGRESADO_ASIGNADO = 'INGRESADO_ASIGNADO',
    CANCELADO = 'CANCELADO',
}

export enum PartRequestType {
    NACIONAL = 'NACIONAL',
    INTERNACIONAL = 'INTERNACIONAL',
}